front/syn_metadata: move UnaryNotOperandKind + const/op value-type helpers out of front/ast (Slice 2)
